+++ frontend/package.json
@@ -33,6 +33,7 @@
     "jest": "^0.1.40",
     "jest-cli": "^0.8.2",
     "react": "^0.14.7",
+    "react-addons-test-utils": "^0.14.7",
     "react-dom": "^0.14.7",
     "react-fontawesome": "^0.3.3",
     "react-router": "^1.0.3",
