+++ frontend/src/js/components/budget-type/BudgetTypeMain.js
@@ -1,0 +1,24 @@
+/**
+ * Created by
+ * Pratish Shrestha <pratishshrestha@lftechnology.com>
+ * on 2/15/16.
+ */
+
+;(function () {
+    'use strict';
+
+    var React = require('react');
+    var BudgetMain = React.createClass({
+
+        render: function () {
+            return (
+                <div id="page-content">
+                    {this.props.children}
+                </div>
+            )
+        }
+    });
+
+    module.exports = BudgetMain;
+
+})();
