+++ frontend/src/js/components/budget-type/BudgetTypeRow.js
@@ -1,0 +1,34 @@
+/**
+ * Created by
+ * Pratish Shrestha <pratishshrestha@lftechnology.com>
+ * on 2/15/16.
+ */
+
+;(function () {
+    var React = require('react');
+    var Link = require('react-router').Link;
+    var BudgetTypeRow = React.createClass({
+        render: function () {
+            return (
+                <tr>
+                    <td>{this.props.budgetType.title}</td>
+                    <td className="text-center">
+                        <div className="btn-group"><Link to={'/budgettypes/edit/'+ this.props.budgetType.id}
+                                                         data-toggle="tooltip"
+                                                         title="Edit"
+                                                         className="btn btn-sm btn-default"><i
+                            className="fa fa-pencil"></i></Link>
+                            <button
+                                className="btn btn-sm btn-default"
+                                onClick={this.props.deleteBudgetType.bind(null, this.props.budgetType.id)}><i
+                                className="glyphicon glyphicon-trash"></i></button>
+                        </div>
+                    </td>
+                </tr>
+            )
+        }
+    });
+
+    module.exports = BudgetTypeRow;
+
+})();
