+++ frontend/src/js/constants/resourceConstant.js
@@ -1,0 +1,14 @@
+/**
+ * Created by
+ * Pratish Shrestha <pratishshrestha@lftechnology.com
+ * on 2/17/16.
+ */
+
+;(function () {
+    const resourceConstant = {
+        BUDGET_TYPES: 'budgettypes'
+    }
+
+    module.exports = resourceConstant;
+
+})();
