+++ frontend/src/js/routes.js
@@ -1,25 +1,27 @@
-'use strict';
+;(function () {
+    'use strict';
 
-var React = require('react'),
-    ReactRouter = require('react-router'),
-    Router = ReactRouter.Router,
-    Route = ReactRouter.Route,
-    createBrowserHistory = require('history/lib/createBrowserHistory'),
-    IndexRoute = ReactRouter.IndexRoute;
+    var React = require('react'),
+        ReactRouter = require('react-router'),
+        Router = ReactRouter.Router,
+        Route = ReactRouter.Route,
+        createBrowserHistory = require('history/lib/createBrowserHistory'),
+        IndexRoute = ReactRouter.IndexRoute;
 
 
-var routes = (
-    <Router history={createBrowserHistory()}>
-        <Route path="/" component={require('./components/App')}>
-            <IndexRoute component={require('./components/dashboard/Dashboard')}/>
-            <Route path="budgettypes" component={require('./components/budget-type/BudgetMain')}>
-                <IndexRoute component={require('./components/budget-type/BudgetList')}/>
-                <Route path="add" component={require('./components/budget-type/BudgetForm')}/>
-                <Route path="edit/:budgetTypeId" component={require('./components/budget-type/BudgetForm')}/>
+    var routes = (
+        <Router history={createBrowserHistory()}>
+            <Route path="/" component={require('./components/App')}>
+                <IndexRoute component={require('./components/dashboard/Dashboard')}/>
+                <Route path="budgettypes" component={require('./components/budget-type/BudgetTypeMain')}>
+                    <IndexRoute component={require('./components/budget-type/BudgetTypeList')}/>
+                    <Route path="new" component={require('./components/budget-type/BudgetTypeForm')}/>
+                    <Route path="edit/:id" component={require('./components/budget-type/BudgetTypeForm')}/>
+                </Route>
             </Route>
-        </Route>
-    </Router>
-);
+        </Router>
+    );
 
-module.exports = routes;
+    module.exports = routes;
 
+})();
